bugfix and raycast

+++ src/gameplay/world/ChunkUtil.hpp
@@ -4,7 +4,7 @@
 
 namespace util::chunk {
 
-glm::vec3 WorldToChunkPos(const glm::ivec3& world_pos);
-glm::vec3 WorldToPosInChunk(const glm::ivec3& world_pos);
+glm::ivec3 WorldToChunkPos(const glm::ivec3& world_pos);
+glm::ivec3 WorldToPosInChunk(const glm::ivec3& world_pos);
 
 }  // namespace util::chunk
